Add effect flag enums Print enums with unknown flags as known + extra Include properties in object printing by default

+++ src/AC2RE.Definitions/FX/CameraFX.cs
@@ -8,7 +8,7 @@ namespace AC2RE.Definitions {
         [Flags]
         public enum PackFlag : ulong {
             NONE = 0,
-            ALL = uint.MaxValue,
+            ALL = ulong.MaxValue,
 
             SHAKE = 1 << 0, // 0x00000001
             LOOKAT = 1 << 1, // 0x00000002
