+++ src/AC2RE.Definitions/Net/NetBlob/NetBlobId.cs
@@ -7,7 +7,7 @@ namespace AC2RE.Definitions {
         [Flags]
         public enum Flag : ulong {
             NONE = 0,
-            ALL = uint.MaxValue,
+            ALL = ulong.MaxValue,
 
             OUT_OF_WORLD = 0x2000000000000000,
             CELL = 0x4000000000000000,
